requirements.txt
openai>=1.0.0
agentops>=0.4.2
python-dotenv>=1.0.0
playwright>=1.30.0
pillow>=9.0.0
fpdf>=1.7.2
